requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
cffi==1.11.5
click==6.7
colorama==0.3.9
coverage
coveralls
cryptography==2.3.1
Flask==1.0.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
pycparser==2.18
PyJWT==1.4.2
pylint==2.1.1
pytest==3.8.0
pytest-cov
pytz==2018.5
PyYAML==3.13
six==1.11.0
typed-ast==1.1.0
# uwsgi
Werkzeug==0.14.1
wrapt==1.10.11
psycopg2
flasgger
certifi
chardet
docopt
Flask-Cors
jsonschema
mistune
requests
urllib3